Assembling the project structure, creating the initial files, and setting up the environment for logic with JavaScript
Performing the initial steps of the project. Creating the basic structure of the application, with the 'index.html' file containing the HTML code of the project, the 'style.css' file containing the application's styles, and the 'assets' folder containing static files such as the page icon. Constructing the entire HTML structure of the page, using the 'phosphor-icons' library to handle project icons. Crafting all project styles in the 'style.css' file and adding the classes to be used by the JavaScript code. Finally, enhancing button accessibility and preparing the environment for the application of business logic through JavaScript.