Starting the development of the application logic, creating the first JavaScript files and the functionality to switch the application between dark mode and light mode.
Continuing the development of the project, starting the construction of the application's business logic. Creating the 'src' folder to contain the project's source code and the 'main.js' file to be the starting point of the application logic. Finally, creating the 'toggle-mode.js' file with the logic responsible for changing the site's theme, toggling between dark and light modes.