Finalizing the application by creating the logic for the counter, adding sounds, and completing the business logic of the site
Concluding the development of the website, creating the files and logic for the counter, and finalizing the business logic of the site. Creating the 'FocusTimer' directory within the 'src' folder to group the JavaScript files responsible for the counter/timer logic and separate these files from the code responsible for the theme-changing logic. Creating, within the 'FocusTimer' folder, the 'index.js' file that contains the 'start' function triggering events and performing the initial update of the counter's minutes and seconds, serving as the starting point for the functionality. Creating the 'state.js' file that stores the initial state of the application through an object. The 'elements.js' file stores variables containing the DOM elements that will be manipulated throughout the application and exports them to other files that will perform this manipulation. The 'events.js' file is responsible for registering all events that will be triggered through the DOM. The 'actions.js' file contains functions responsible for the control logic and will be triggered through events. The 'timer.js' file stores the logic of the counter itself, containing the 'updateDisplay' function responsible for updating the minutes and seconds of the counter on the screen and the 'countdown' function responsible for the timer countdown logic. Finally, adding sound files to the 'assets' folder and creating the 'sounds.js' file that transforms these sounds into JavaScript objects and stores them in variables to be used in other files that will play these sounds